Replace crap desktop notifications with crap GTK+ dialogs
mcatanzaro:  I am going to remove the "display events in notification
area" crap
mcatanzaro:  Because the notification goes away and then you can never
see it again or join the room
mcatanzaro:  It's BS
mcatanzaro:  If you toggle it off, you get a GtkDialog from nowhere
instead, which is also crap but better!

To be clear: it's totally unacceptable to display GtkDialogs without a
parent window, and this patch makes us do that by default. But I don't
plan to spend more effort on this myself, and it's a lot better than
giving users a two second window to rush to click the buttons in the
notification before it disappears. Better hope no other notification
gets sent during that time, and you have fast reflexes, else you're not
gonna be able to join the chat you're invited to....