roster-window: Fix the GActions not working
Underscores are not allowed in GAction names. This led to a pile of
warnings during startup and ultimately none of these actions working.
Fix by substituting the underscores with hyphens, an allowed character.

We can also omit the NULLs in the GActionEntries, as they are implied:
struct fields not explicitly initialised are zero-initialised anyway.