tests: Retain the PEM formatted root certificate
In the subsequent commit, we will use GIO's GnuTLS backend to verify
the chain of TLS certificates instead of Gcr. This means that a
GckModule can no longer be used to feed our mock root certificates via
a PKCS#11 module. Instead we will have to create a mock GTlsDatabase,
and that needs the PEM formatted root certificate.

We will continue using the GckModule for pinned certificates.

https://bugzilla.gnome.org/show_bug.cgi?id=780160